+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 ## [Unreleased]
 
+## 0.5.3
+
+### Features
+
+- Added default package 'stoyco_subscription' name in assets.gen for assets tokens
+- replace SvgPicture.asset with generated asset references in multiple components
+
+## [Unreleased]
+
 ## 0.5.2
 
 ### Features